64 bit work except usb

+++ fb/video_fb.h
@@ -57,7 +57,7 @@ typedef struct {
 	unsigned int dprBase;
 	unsigned int vprBase;
 	unsigned int cprBase;
-	unsigned int frameAdrs;
+	unsigned long frameAdrs;
 	unsigned int memSize;
 	unsigned int mode;
 	unsigned int gdfIndex;
