+++ include/pflash.h
@@ -62,7 +62,7 @@ struct fl_device {
  *  Structure describing targets FLASH memory map.
  */
 struct fl_map {
-	u_int32_t fl_map_base;	/* Start of flash area in physical map */
+	unsigned long fl_map_base;	/* Start of flash area in physical map */
 	u_int32_t fl_map_size;	/* Size of flash area in physical map */
 	int	fl_map_width;	/* Number of bytes to program in one cycle */
 	int	fl_map_chips;	/* Number of chips to operate in one cycle */
